+++ .pre-commit-config.yaml
@@ -95,6 +95,7 @@ repos:
         exclude: |
             (?x)^(
                 starboard/android/apk/app/src/main/java/dev/cobalt/storage/StorageProto.java|
+                cobalt/android/apk/app/src/org/chromium/content_shell/Util.java|
                 cobalt/android/apk/app/src/org/chromium/content_shell/Shell.java|
                 cobalt/android/apk/app/src/org/chromium/content_shell/ShellManager.java
             )
